+++ src/main/java/com/troblecodings/signals/guis/GuiPathwayRequester.java
@@ -54,8 +54,8 @@ public class GuiPathwayRequester extends GuiBase {
                         return "Disabled";
                     return container.validStarts.get(e).toString();
                 });
-        final IIntegerable<String> end = SizeIntegerables.of("EndPoint",
-                container.validStarts.size(), e -> {
+        final IIntegerable<String> end = SizeIntegerables.of("EndPoint", container.validEnds.size(),
+                e -> {
                     if (e == -1)
                         return "Disabled";
                     return container.validEnds.get(e).toString();