+++ src/main/java/com/troblecodings/signals/guis/GuiSignalController.java
@@ -1,6 +1,5 @@
 package com.troblecodings.signals.guis;
 
-import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -49,7 +48,6 @@ public class GuiSignalController extends GuiBase {
 
     private final ContainerSignalController controller;
     private final UIEntity lowerEntity = new UIEntity();
-    private final List<UIPropertyEnumHolder> holders = new ArrayList<>();
     private boolean loaded = false;
     private final PlayerEntity player;
     private EnumMode currentMode;
@@ -300,10 +298,9 @@ public class GuiSignalController extends GuiBase {
         lowerEntity.add(previewSidebar.get());
         lowerEntity.add(new UIBox(UIBox.HBOX, 1));
 
-        holders.clear();
         ClientSignalStateHandler.getClientStates(new StateInfo(mc.level, controller.getPos()))
-        .forEach((property, value) -> previewSidebar.addToRenderNormal(property,
-                property.getParent().getIDFromValue(value)));
+                .forEach((property, value) -> previewSidebar.addToRenderNormal(property,
+                        property.getParent().getIDFromValue(value)));
         final Map<SEProperty, String> map = this.controller.getProperties();
         if (map == null)
             return;
@@ -317,7 +314,6 @@ public class GuiSignalController extends GuiBase {
                     previewSidebar.update(controller.getSignal());
                 }
             }, index));
-            holders.add(new UIPropertyEnumHolder(property, enumarable));
             previewSidebar.addToRenderNormal(property, index);
 
         });