+++ src/main/java/com/troblecodings/signals/signalbox/SignalBoxNode.java
@@ -17,6 +17,7 @@ import com.google.common.collect.ImmutableSet;
 import com.troblecodings.core.NBTWrapper;
 import com.troblecodings.core.ReadBuffer;
 import com.troblecodings.core.WriteBuffer;
+import com.troblecodings.signals.core.TrainNumber;
 import com.troblecodings.signals.enums.EnumGuiMode;
 import com.troblecodings.signals.enums.PathType;
 import com.troblecodings.signals.signalbox.debug.SignalBoxFactory;
@@ -38,6 +39,7 @@ public class SignalBoxNode implements INetworkSavable, Iterable<ModeSet> {
     private final Point point;
     private boolean isAutoPoint = false;
     private String customText = "";
+    private TrainNumber trainNumber = TrainNumber.DEFAULT;
 
     public SignalBoxNode() {
         this(new Point());
@@ -77,7 +79,7 @@ public class SignalBoxNode implements INetworkSavable, Iterable<ModeSet> {
         manuellEnabledOutputs.clear();
         return returnList;
     }
-    
+
     public List<ModeSet> getManuellEnabledOutputs() {
         return ImmutableList.copyOf(manuellEnabledOutputs);
     }
@@ -92,6 +94,18 @@ public class SignalBoxNode implements INetworkSavable, Iterable<ModeSet> {
 
     public boolean isAutoPoint() {
         return isAutoPoint;
+    }
+
+    public void setTrainNumber(final TrainNumber number) {
+        this.trainNumber = number;
+    }
+
+    public void removeTrainNumber() {
+        this.trainNumber = TrainNumber.DEFAULT;
+    }
+
+    public TrainNumber getTrainNumber() {
+        return trainNumber == null ? TrainNumber.DEFAULT : trainNumber;
     }
 
     public void post() {
@@ -199,6 +213,8 @@ public class SignalBoxNode implements INetworkSavable, Iterable<ModeSet> {
         this.point.write(compound);
         compound.putBoolean(IS_AUTO_POINT, isAutoPoint);
         compound.putString(CUSTOM_NAME, customText);
+        if (this.trainNumber != null)
+            trainNumber.writeTag(compound);
     }
 
     @Override
@@ -217,6 +233,7 @@ public class SignalBoxNode implements INetworkSavable, Iterable<ModeSet> {
         this.point.read(compound);
         this.isAutoPoint = compound.getBoolean(IS_AUTO_POINT);
         this.customText = compound.getString(CUSTOM_NAME);
+        this.trainNumber = TrainNumber.of(compound);
         post();
     }
 
@@ -302,7 +319,7 @@ public class SignalBoxNode implements INetworkSavable, Iterable<ModeSet> {
         return this.possibleModes.keySet().stream()
                 .anyMatch(modeSet -> VALID_MODES.contains(modeSet.mode));
     }
-    
+
     public boolean isValidEnd() {
         return this.possibleModes.keySet().stream()
                 .anyMatch(modeSet -> VALID_MODES.contains(modeSet.mode)
@@ -346,12 +363,9 @@ public class SignalBoxNode implements INetworkSavable, Iterable<ModeSet> {
             if (!manuellEnabledOutputs.contains(modeSet))
                 manuellEnabledOutputs.add(modeSet);
         }
-        this.isAutoPoint = buffer.getByte() == 1 ? true : false;
-        final int nameSize = buffer.getByteToUnsignedInt();
-        final byte[] array = new byte[nameSize];
-        for (int i = 0; i < nameSize; i++)
-            array[i] = buffer.getByte();
-        this.customText = new String(array);
+        this.isAutoPoint = buffer.getBoolean();
+        this.customText = buffer.getString();
+        this.trainNumber = TrainNumber.of(buffer);
         post();
     }
 
@@ -372,12 +386,9 @@ public class SignalBoxNode implements INetworkSavable, Iterable<ModeSet> {
             if (!manuellEnabledOutputs.contains(modeSet))
                 manuellEnabledOutputs.add(modeSet);
         }
-        this.isAutoPoint = buffer.getByte() == 1 ? true : false;
-        final int nameSize = buffer.getByteToUnsignedInt();
-        final byte[] array = new byte[nameSize];
-        for (int i = 0; i < nameSize; i++)
-            array[i] = buffer.getByte();
-        this.customText = new String(array);
+        this.isAutoPoint = buffer.getBoolean();
+        this.customText = buffer.getString();
+        this.trainNumber = TrainNumber.of(buffer);
         post();
     }
 
@@ -390,11 +401,13 @@ public class SignalBoxNode implements INetworkSavable, Iterable<ModeSet> {
         });
         buffer.putByte((byte) manuellEnabledOutputs.size());
         manuellEnabledOutputs.forEach(mode -> mode.writeNetwork(buffer));
-        buffer.putByte((byte) (isAutoPoint ? 1 : 0));
-        final byte[] array = customText.getBytes();
-        buffer.putByte((byte) array.length);
-        for (final byte b : array)
-            buffer.putByte(b);
+        buffer.putBoolean(isAutoPoint);
+        buffer.putString(customText);
+        if (trainNumber != null) {
+            trainNumber.writeNetwork(buffer);
+        } else {
+            TrainNumber.DEFAULT.writeNetwork(buffer);
+        }
     }
 
     public void writeUpdateNetwork(final WriteBuffer buffer) {
@@ -411,11 +424,13 @@ public class SignalBoxNode implements INetworkSavable, Iterable<ModeSet> {
             }
         });
         buffer.putByte((byte) 0);
-        buffer.putByte((byte) (isAutoPoint ? 1 : 0));
-        final byte[] array = customText.getBytes();
-        buffer.putByte((byte) array.length);
-        for (final byte b : array)
-            buffer.putByte(b);
+        buffer.putBoolean(isAutoPoint);
+        buffer.putString(customText);
+        if (trainNumber != null) {
+            trainNumber.writeNetwork(buffer);
+        } else {
+            TrainNumber.DEFAULT.writeNetwork(buffer);
+        }
     }
 
     public String getCustomText() {